# src/MarqSpec.TradingCopilot.Client/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Trading Co-Pilot API",
    "description": "The single-operator futures trading co-pilot BFF (R-10). Generated from the live minimal-API routes (gh#604) — the source of truth the README links to, replacing the hand-kept table.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
...
    "/accounts/{id}/stage": {
      "put": {
        "tags": ["Accounts"],
        "summary": "Set a manual stage override on an account.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetStageOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      },
      "delete": {
        "tags": ["Accounts"],
        "summary": "Clear the manual stage override on an account.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      }
    },
...
    "/api/news": {
      "get": {
        "tags": ["News"],
        "summary": "The personalized feed: recent news re-ranked by the operator's decayed salience profile.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": ["integer", "string"],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      }
    },
    "/api/news/feedback": {
      "put": {
        "tags": ["News"],
        "summary": "Stars or mutes a news item — an upsert: re-rating replaces, so an item carries at most one feedback.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      },
      "delete": {
        "tags": ["News"],
        "summary": "Un-stars / un-mutes an item — the personalization round-trips, so its salience returns to base.",
        "parameters": [
          {
            "name": "dedupKey",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountStage": {
        "type": "integer",
        "description": "Where an account sits in a firm's programme. Read from what the venue reports — for ProjectX that means the\naccount <b>name</b>, which really does encode size, stage, and status."
      },
...
      "NewsFeedbackRequest": {
        "required": ["dedupKey", "kind"],
        "type": "object",
        "properties": {
          "dedupKey": {
            "type": "string",
            "description": "The rated item's dedup key (its NewsRecord primary key)."
          },
          "kind": {
            "description": "Star or mute (never SoftSignalKind.Unknown).",
            "$ref": "#/components/schemas/SoftSignalKind"
          }
        },
        "description": "A star/mute request on a news item (gh#27). The dedup key rides in the body — it is a URL, not path-safe."
      },
...
      "SetStageOverrideRequest": {
        "required": ["stage"],
        "type": "object",
        "properties": {
          "stage": {
            "description": "The declared stage. `Unknown` is refused — clearing the override says that.",
            "$ref": "#/components/schemas/AccountStage"
          }
        },
        "description": "The request to declare an account's stage, overriding the conservative name-resolver (gh#76)."
      },
...
      "SoftSignalKind": {
        "type": "integer",
        "description": "The kind of per-user feedback on a soft-signal / news item (gh#27, ADR-0014). A <b>refusable zero</b>:\nSoftSignalKind.Unknown is never a real kind — feedback must declare its intent — refused at construction and by a\nDB check. One entity (`SoftSignalFeedback`) unifies all kinds; <b>star</b> and <b>mute</b> ship now, and the\n👍/👎 <b>sentiment</b> (direction) axis is deferred (R-2) and will extend this enum without reshaping the store."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT issued by POST /auth/login, sent as: Authorization: Bearer {token}.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}